pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = ["Luciano2078 <luciano.pereiraaristimunho@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^6.2.5"
factory-boy = "^3.2.1"
Django = "^4.0"
django-rest-framework = "^0.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
